feat(session): refactor for template
