+++ src/test/jandedobbeleer-palette.omp.json
@@ -43,8 +43,7 @@
           "leading_diamond": "",
           "trailing_diamond": "\uE0B0",
           "properties": {
-            "postfix": " ",
-            "display_host": false
+            "template": "{{ .UserName }}"
           }
         },
         {
