+++ src/test/jandedobbeleer.omp.json
@@ -12,8 +12,7 @@
           "leading_diamond": "",
           "trailing_diamond": "\uE0B0",
           "properties": {
-            "postfix": " ",
-            "display_host": false
+            "template": "{{ .UserName }}"
           }
         },
         {
