+++ themes/M365Princess.omp.json
@@ -15,7 +15,7 @@
           "trailing_diamond": "",
           "properties": {
             "prefix": "",
-            "display_host": false
+            "template": "{{ .UserName }}"
           }
         },
         {
