+++ themes/aliens.omp.json
@@ -11,7 +11,10 @@
           "foreground": "#ffffff",
           "background": "#61AFEF",
           "leading_diamond": "\uE0B6",
-          "trailing_diamond": "\uE0B0"
+          "trailing_diamond": "\uE0B0",
+          "properties": {
+            "template": "{{ .UserName }}@{{ .ComputerName }}"
+          }
         },
         {
           "type": "path",
