+++ themes/powerlevel10k_classic.omp.json
@@ -64,7 +64,8 @@
           "background": "#546E7A",
           "leading_diamond": "\uE0B2",
           "properties": {
-            "postfix": " <#26C6DA>\uE0B3</> "
+            "postfix": " <#26C6DA>\uE0B3</> ",
+            "template": "{{ .UserName }}@{{ .ComputerName }}"
           }
         },
         {
